refactor: move classes to more accurate packages
The exception class and the interface just make more sense to be in the
transformers package.
