+++ engine/src/main/java/com/redhat/cloud/notifications/exports/transformers/ResultsTransformer.java
@@ -1,4 +1,4 @@
-package com.redhat.cloud.notifications.exports;
+package com.redhat.cloud.notifications.exports.transformers;
 
 import java.util.List;
