+++ engine/src/main/java/com/redhat/cloud/notifications/exports/transformers/event/CSVEventTransformer.java
@@ -1,7 +1,7 @@
 package com.redhat.cloud.notifications.exports.transformers.event;
 
-import com.redhat.cloud.notifications.exports.ResultsTransformer;
-import com.redhat.cloud.notifications.exports.TransformationException;
+import com.redhat.cloud.notifications.exports.transformers.ResultsTransformer;
+import com.redhat.cloud.notifications.exports.transformers.TransformationException;
 import com.redhat.cloud.notifications.models.Event;
 import org.apache.commons.csv.CSVFormat;
 import org.apache.commons.csv.CSVPrinter;
