+++ engine/src/test/java/com/redhat/cloud/notifications/exports/transformers/event/JSONEventTransformerTest.java
@@ -1,7 +1,7 @@
 package com.redhat.cloud.notifications.exports.transformers.event;
 
-import com.redhat.cloud.notifications.exports.ResultsTransformer;
-import com.redhat.cloud.notifications.exports.TransformationException;
+import com.redhat.cloud.notifications.exports.transformers.ResultsTransformer;
+import com.redhat.cloud.notifications.exports.transformers.TransformationException;
 import com.redhat.cloud.notifications.exports.transformers.TransformersHelpers;
 import com.redhat.cloud.notifications.models.Event;
 import io.vertx.core.json.JsonArray;
